fix(tsconfig): use find-root to find the nearest tsconfig.json
app-root-path relies on its own install location to determine the root of the project.
however, when using yarn workspaces, this assumption is incorrect.
this results in either not finding a tsconfig.json or finding a tsconfig.json that has been extended.
find-root allows passing a starting location and predicate and searches recursively up.