Adding ability to create custom transformers.
This attempts to be fully backwards compatible with all previous versions. Now, we export a createTransformer function from said file to create the transform function. The base file still exports this transformer in the same way it did before.

This allows implementers to create their own transformers that essentially just set configuration. The only configuration option I could think to set right now is to set a path for the TypeScript config, tsConfigPath.